Renamed the expected files etc. accordingly
