+++ mipengine/algorithms/anova_oneway.json
