+++ tests/algorithm_validation_tests/test_anova_oneway_validation.py
@@ -7,7 +7,7 @@ import requests
 
 from tests.prod_env_tests import algorithms_url
 
-expected_file = Path(__file__).parent / "expected" / "one_way_anova_expected.json"
+expected_file = Path(__file__).parent / "expected" / "anova_oneway_expected.json"
 
 
 def anova_one_way_request(input):
