+++ tests/testcase_generators/pearson_testcase_generator.py
@@ -6,13 +6,13 @@ from scipy import stats
 
 from tests.testcase_generators.testcase_generator import TestCaseGenerator
 
-SPECS_PATH = Path("mipengine", "algorithms", "pearson.json")
+SPECS_PATH = Path("mipengine", "algorithms", "pearson_correlation.json")
 EXPECTED_PATH = Path(
     "tests",
     "prod_env_tests",
     "algorithm_validation_tests",
     "expected",
-    "pearson_expected.json",
+    "pearson_correlation_expected.json",
 )
